reloc: fold marker-map loader into priors, ultra-light comments, restore leshy's relocalize comments
