+++ dimos/mapping/relocalization/module.py
@@ -34,6 +34,7 @@ from dimos.mapping.relocalization.priors import (
     RansacPrior,
     RansacPriorConfig,
     RelocPrior,
+    load_marker_map,
     relocalize_with_priors as _relocalize_with_priors,
 )
 from dimos.mapping.relocalization.relocalize import (
@@ -49,7 +50,6 @@ from dimos.msgs.sensor_msgs.PointCloud2 import PointCloud2
 from dimos.msgs.vision_msgs.Detection3D import Detection3D
 from dimos.msgs.vision_msgs.Detection3DArray import Detection3DArray
 from dimos.perception.fiducial.apriltag_aggregation import matrix_from_pose7
-from dimos.perception.fiducial.fiducial_relocalization import load_marker_map
 from dimos.utils.data import resolve_named_path
 from dimos.utils.logging_config import setup_logger
 from dimos.utils.reactive import backpressure
@@ -64,15 +64,11 @@ MAP_SUFFIX = ".pc2.lcm"
 # Suffixes load_marker_map parses; a bare name gets the .json default (see _start_fiducial_prior).
 MARKER_MAP_SUFFIXES = (".json", ".yaml", ".yml")
 SKIP_LOG_INTERVAL_S = 5.0  # s; throttle relocalize-skip warnings so a starved feed can't spam
-# Jump guard, TRACKING only: largest per-second step from the previous accepted fix.
-# T is the world->map alignment, so between accepts it moves only by the drift the fix
-# corrects; this refuses the mirror-flipped fix that lands a room over. Floored at one
-# second's worth below so back-to-back priors keep a real budget.
+# Jump guard, TRACKING only: largest per-second step from the previous accepted fix. T is world->map, so between accepts it moves only by the drift the fix corrects; this refuses the mirror-flipped fix that lands a room over. Floored at one second's worth below so back-to-back priors keep a real budget.
 MAX_JUMP_M_PER_S = 5.0  # m/s; 10 m at the 2 s RANSAC interval
 MAX_JUMP_YAW_DEG_PER_S = 45.0  # deg/s; 90 deg at the 2 s RANSAC interval
 
-# Keys that moved onto the prior entries. extra="forbid" already raises on them, but
-# the redirect message says where the knob went instead of "extra inputs not permitted".
+# Keys that moved onto the prior entries. extra="forbid" already raises, but the redirect message says where the knob went instead of "extra inputs not permitted".
 _MOVED_TO_PRIORS = {
     "reloc_interval_s": (
         "reloc_interval_s moved onto the ransac prior entry; set "
@@ -96,20 +92,16 @@ class Config(ModuleConfig):
     map_file: str | None = (
         None  # e.g. `-o relocalizationmodule.map_file=go2_hongkong_office_twopass_map`
     )
-    # Operator -o override for the fiducial marker map (a dotted -o can't index the
-    # priors list); wins over the blueprint entry in _start_fiducial_prior. None -> use the entry's own.
+    # Operator -o override for the fiducial marker map (a dotted -o can't index the priors list); wins over the blueprint entry in _start_fiducial_prior. None -> use the entry's own.
     marker_map_file: str | None = None
     publish_loaded_map: bool = False
     # fitness_threshold and min_local_points moved onto the prior entries; see _MOVED_TO_PRIORS.
-    # Max z-axis tilt (deg) a candidate may keep at the judge's gravity gate;
-    # matches relocalize.py's GRAVITY_TILT_MAX_DEG, so unchanged unless overridden.
+    # Max z-axis tilt (deg) a candidate may keep at the judge's gravity gate; matches relocalize.py's GRAVITY_TILT_MAX_DEG, so unchanged unless overridden.
     gravity_tilt_max_deg: float = Field(default=10.0, ge=0.0)
     use_carving: bool = True
-    # False (default): one line per accepted fix plus throttled warnings. True (`--eval`):
-    # each accept also logs its published pose, timing and point count.
+    # False (default): one line per accepted fix plus throttled warnings. True (`--eval`): each accept also logs its published pose, timing and point count.
     verbose_eval_logging: bool = False
-    # The prior pool: each entry a toggleable candidate proposer (RANSAC polled, fiducial event-driven).
-    # REQUIRED, no default -- a blueprint must state it, so there is no silent reloc behavior.
+    # The prior pool: each entry a toggleable candidate proposer (RANSAC polled, fiducial event-driven). REQUIRED, no default -- a blueprint must state it, so there is no silent reloc behavior.
     priors: list[PriorConfig]
 
     @model_validator(mode="before")
@@ -133,43 +125,32 @@ class RelocalizationModule(Module):
         super().__init__(**kwargs)
         self._premap: PointCloud2 | None = None
         self._last_skip_log = 0.0
-        # Per-source accept/reject tally, filled only under verbose_eval_logging (--eval)
-        # and rendered once at stop(). The module already owns the accept/reject data.
+        # Per-source accept/reject tally, filled only under verbose_eval_logging (--eval) and rendered once at stop(). The module already owns the accept/reject data.
         self._eval_tally: dict[str, SourceTally] = {}
         self._world_to_map: Subject[Transform | None] = Subject()
-        # Prior objects are built ONCE, not per frame: the fiducial holds pending-fix state
-        # across bursts that a fresh instance would reset. The RANSAC prior is a pure source;
-        # the module owns its poll timer (below).
+        # Prior objects are built ONCE, not per frame: the fiducial holds pending-fix state across bursts that a fresh instance would reset. The RANSAC prior is a pure source; the module owns its poll timer (below).
         ransac_entry = next(
             (p for p in self.config.priors if isinstance(p, RansacPriorConfig)),
             RansacPriorConfig(),
         )
         self._ransac_prior = RansacPrior()
-        # Per-source accept gate {Candidate.source: min wall fitness}. Each fire pools one
-        # source, so gating the winner by its own source's bar IS per-source. Key = the
-        # entry's `type` discriminator, which equals the source its prior proposes.
+        # Per-source accept gate {Candidate.source: min wall fitness}. Each fire pools one source, so gating the winner by its own source's bar IS per-source. Key = the entry's `type` discriminator, which equals the source its prior proposes.
         self._accept_threshold: dict[str, float] = {
             p.type: p.fitness_threshold for p in self.config.priors
         }
-        # RANSAC poll timer (the prior is a pure source): its interval and last fire.
-        # None == never fired, so the first dense-enough frame relocalizes immediately.
+        # RANSAC poll timer (the prior is a pure source): its interval and last fire. None == never fired, so the first dense-enough frame relocalizes immediately.
         self._ransac_interval_s = ransac_entry.interval_s
         self._last_ransac_fired_s: float | None = None
-        # RANSAC-only point floor (a fiducial burst fires regardless); read from the
-        # ransac entry, or the default when no ransac entry is enabled (then unused).
+        # RANSAC-only point floor (a fiducial burst fires regardless); read from the ransac entry, or the default when no ransac entry is enabled (then unused).
         self._ransac_min_local_points = ransac_entry.min_local_points
         # Built in start() once the marker map is loaded (needs map_T_marker).
         self._fiducial_prior: FiducialPrior | None = None
-        # Previous ACCEPTED fix (map_T_world) and when, for the gross-jump guard.
-        # None until the first accept, which is what leaves acquisition unguarded.
+        # Previous ACCEPTED fix (map_T_world) and when, for the gross-jump guard. None until the first accept, which is what leaves acquisition unguarded.
         self._last_fix_map_T_world: np.ndarray | None = None
         self._last_fix_ts_s = 0.0
-        # Latest global_map, so a completed tag burst is judged the instant it lands. Sound
-        # because the stream ACCUMULATES in the world frame: the previous cloud scores wall
-        # fitness as well as the newest, and the tag candidate needs no lidar at all.
+        # Latest global_map, so a completed tag burst is judged the instant it lands. Sound because the stream ACCUMULATES in the world frame: the previous cloud scores wall fitness as well as the newest, and the tag candidate needs no lidar at all.
         self._last_local_map: PointCloud2 | None = None
-        # One monotonic timebase for every prior trigger, injectable so a test can
-        # drive the triggers without sleeping (FiducialPrior takes the same seam).
+        # One monotonic timebase for every prior trigger, injectable so a test can drive the triggers without sleeping (FiducialPrior takes the same seam).
         self._now_fn: Callable[[], float] = time.monotonic
 
     @rpc
@@ -184,10 +165,7 @@ class RelocalizationModule(Module):
         self._premap = PointCloud2.lcm_decode(path.read_bytes())
         self._premap.frame_id = FRAME_MAP
 
-        # No throttle here: the priors own the cadence. EVERY frame reaches _on_local_map,
-        # which asks each prior whether it wants a fire. Trigger accounting runs on the
-        # backpressure worker thread only -- a frame backpressure coalesces away never
-        # consumes a trigger, so a burst edge cannot be dropped by a slow solve.
+        # No throttle here: the priors own the cadence. EVERY frame reaches _on_local_map, which asks each prior whether it wants a fire. Trigger accounting runs on the backpressure worker thread only -- a frame backpressure coalesces away never consumes a trigger, so a burst edge cannot be dropped by a slow solve.
         self.register_disposable(
             backpressure(
                 self.global_map.observable()  # type: ignore[no-untyped-call]
@@ -219,8 +197,7 @@ class RelocalizationModule(Module):
 
     @rpc
     def stop(self) -> None:
-        # Emit the per-source accept/reject table once, at shutdown, from the tally the
-        # fire path filled -- the module owns the data, so no log parsing.
+        # Emit the per-source accept/reject table once, at shutdown, from the tally the fire path filled -- the module owns the data, so no log parsing.
         if self.config.verbose_eval_logging and self._eval_tally:
             logger.info("relocalize eval summary", table=format_eval_summary(self._eval_tally))
         super().stop()
@@ -244,8 +221,7 @@ class RelocalizationModule(Module):
                 "relocalize: fiducial prior enabled but no marker_map_file; fiducial prior disabled"
             )
             return
-        # A bare name gets the .json default; a name that already states its format keeps it
-        # (resolve_named_path would otherwise look up "<survey>.yaml.json", which nothing writes).
+        # A bare name gets the .json default; a name that already states its format keeps it (resolve_named_path would otherwise look up "<survey>.yaml.json", which nothing writes).
         suffix = "" if Path(marker_map_file).suffix in MARKER_MAP_SUFFIXES else ".json"
         marker_map = {
             marker_id: transform.to_matrix()
@@ -275,8 +251,7 @@ class RelocalizationModule(Module):
             self._last_skip_log = now
 
     def _maybe_log_wall_skip(self, n_pts: int, reason: str) -> None:
-        """Throttled warning that the judge refused a fire for too little wall evidence.
-        Shares the RANSAC-skip throttle timer -- both are 'cloud too sparse this fire'."""
+        """Throttled warning that the judge refused a fire for too little wall evidence; shares the RANSAC-skip throttle timer -- both are 'cloud too sparse this fire'."""
         now = time.monotonic()
         if now - self._last_skip_log > SKIP_LOG_INTERVAL_S:
             logger.warning(
@@ -310,9 +285,7 @@ class RelocalizationModule(Module):
                 )
             )
             self._fiducial_prior.observe(marker_id, world_T_marker_aggregated)
-        # Fire HERE, not at the next cloud: the tag candidate is composed from the marker
-        # alone, so waiting on lidar is dead time on acquisition. No cloud yet -> pending
-        # stays and _on_local_map fires it on the first frame.
+        # Fire HERE, not at the next cloud: the tag candidate is composed from the marker alone, so waiting on lidar is dead time on acquisition. No cloud yet -> pending stays and _on_local_map fires it on the first frame.
         cloud = self._last_local_map
         if cloud is not None and self._fiducial_prior.has_pending:
             self._fire(self._fiducial_prior, cloud)
@@ -323,8 +296,7 @@ class RelocalizationModule(Module):
 
     @staticmethod
     def _marker_id_from_detection(detection: Detection3D) -> int | None:
-        """Marker id from a wire Detection3D: the ``id`` field, else the numeric
-        tail of a ``DICT:id`` class label (the detector's marker_label encoding)."""
+        """Marker id from a wire Detection3D: the ``id`` field, else the numeric tail of a ``DICT:id`` class label (the detector's marker_label encoding)."""
         raw = str(getattr(detection, "id", "")).strip()
         if raw.isdigit():
             return int(raw)
@@ -352,8 +324,7 @@ class RelocalizationModule(Module):
         """Poll the RANSAC prior on the module's timer; on cold start fire a pending fiducial fix."""
         self._last_local_map = msg  # what a burst-triggered fire judges against
         now = self._now_fn()
-        # _ransac_prior is always constructed, so gate it on the enabled set (the toggle);
-        # _fiducial_prior is None unless its enabled entry loaded a marker map.
+        # _ransac_prior is always constructed, so gate it on the enabled set (the toggle); _fiducial_prior is None unless its enabled entry loaded a marker map.
         enabled = self._enabled_prior_objects()
 
         if self._ransac_prior in enabled:
@@ -381,9 +352,7 @@ class RelocalizationModule(Module):
             return None  # every prior disabled (or fiducial-only with no marker map yet)
         t0 = time.monotonic()
         winning_source: str | None = None
-        # Plain relocalize() is reserved for the lidar-only MODULE, where the SOLVE is the
-        # pre-prior path bit-for-bit and the accept line carries no source=. Once a tag prior
-        # is configured every fire goes through the judge instead, so both sources are counted.
+        # Plain relocalize() is reserved for the lidar-only MODULE, where the SOLVE is the pre-prior path bit-for-bit and the accept line carries no source=. Once a tag prior is configured every fire goes through the judge instead, so both sources are counted.
         solo_prior_module = len(self._enabled_prior_objects()) == 1
         try:
             if solo_prior_module and len(priors) == 1 and isinstance(priors[0], RansacPrior):
@@ -400,19 +369,15 @@ class RelocalizationModule(Module):
                     gravity_tilt_max_deg=self.config.gravity_tilt_max_deg,
                 )
         except EmptyProposalError:
-            # Double-fire race: the other thread drained the pending tag fix first, so this
-            # cycle judges an empty pool. The winner already published it -- benign no-op.
+            # Double-fire race: the other thread drained the pending tag fix first, so this cycle judges an empty pool. The winner already published it -- benign no-op.
             logger.debug("relocalize: no candidates this cycle")
             return None
         except InsufficientWallEvidenceError as e:
-            # Too few wall points to judge the pool (a sparse acquisition cloud a tag burst
-            # is allowed to fire on): drop the unjudged fix -- a pose scored against <100
-            # walls isn't trustworthy, and the tag is re-seen as the robot moves into structure.
+            # Too few wall points to judge the pool (a sparse acquisition cloud a tag burst is allowed to fire on): drop the unjudged fix -- a pose scored against <100 walls isn't trustworthy, and the tag is re-seen as the robot moves into structure.
             self._maybe_log_wall_skip(len(msg), str(e))
             return None
         except NoUprightCandidateError:
-            # Every candidate tilted past the gravity gate: a real rejection, so the fix is
-            # consumed (re-judging a tilted pose only re-rejects it). Throttled warn, shares the skip window.
+            # Every candidate tilted past the gravity gate: a real rejection, so the fix is consumed (re-judging a tilted pose only re-rejects it). Throttled warn, shares the skip window.
             now = time.monotonic()
             if now - self._last_skip_log > SKIP_LOG_INTERVAL_S:
                 logger.warning("relocalize rejected: all candidates tilted past gravity gate")
@@ -423,20 +388,16 @@ class RelocalizationModule(Module):
             return None
         dt = time.monotonic() - t0
         n_pts = len(msg)
-        # source= emission: ACCEPT names the prior that WON (omitted on the single-source
-        # path, which the parsers read as ransac); REJECT names the prior whose BAR refused
-        # the fix, so --eval can bucket rejects by source instead of "unknown".
+        # source= emission: ACCEPT names the prior that WON (omitted on the single-source path, which the parsers read as ransac); REJECT names the prior whose BAR refused the fix, so --eval can bucket rejects by source instead of "unknown".
         source_kw: dict[str, str] = {} if winning_source is None else {"source": winning_source}
 
-        # The winner is gated on ITS OWN source's bar. winning_source is None only on the
-        # plain (ransac-only) path -> resolve to the ransac entry.
+        # The winner is gated on ITS OWN source's bar. winning_source is None only on the plain (ransac-only) path -> resolve to the ransac entry.
         gate_source = winning_source if winning_source is not None else RansacPrior.name
         threshold = self._accept_threshold[gate_source]
         if fitness < threshold:
             if self.config.verbose_eval_logging:
                 self._eval_tally.setdefault(gate_source, SourceTally()).rejects += 1
-            # threshold= IS the reason: this is the only reject path, and the
-            # operator's next move is to compare the two numbers.
+            # threshold= IS the reason: this is the only reject path, and the operator's next move is to compare the two numbers.
             extra: dict[str, Any] = (
                 {"time_cost_s": round(dt, 1), "n_pts": n_pts}
                 if self.config.verbose_eval_logging
@@ -451,20 +412,15 @@ class RelocalizationModule(Module):
             )
             return None
 
-        # Jump guard, TRACKING only: a fix stepping further from the previous one than the
-        # budgets above allow is a mis-localization, not a drift correction. No previous fix
-        # means acquisition, which is never blocked. Both terms are on map_T_world.
+        # Jump guard, TRACKING only: a fix stepping further from the previous one than the budgets above allow is a mis-localization, not a drift correction. No previous fix means acquisition, which is never blocked. Both terms are on map_T_world.
         now_s = self._now_fn()
         if self._last_fix_map_T_world is not None:
             dt_s = now_s - self._last_fix_ts_s
-            # Floored at one second's worth: two priors fire back to back may disagree by a
-            # normal correction, and a budget shrinking with the gap would refuse exactly the
-            # cross-source fix the fiducial preset exists to publish.
+            # Floored at one second's worth: two priors fire back to back may disagree by a normal correction, and a budget shrinking with the gap would refuse exactly the cross-source fix the fiducial preset exists to publish.
             budget_s = max(dt_s, 1.0)
             jump_m = float(np.linalg.norm(T[:3, 3] - self._last_fix_map_T_world[:3, 3]))
             rot_rel = self._last_fix_map_T_world[:3, :3].T @ T[:3, :3]
-            # Yaw is the whole story: the judge's gravity gate already holds tilt under
-            # GRAVITY_TILT_MAX_DEG, so a flip between two accepted fixes is about z.
+            # Yaw is the whole story: the judge's gravity gate already holds tilt under GRAVITY_TILT_MAX_DEG, so a flip between two accepted fixes is about z.
             yaw_deg = abs(float(np.degrees(np.arctan2(rot_rel[1, 0], rot_rel[0, 0]))))
             if jump_m > MAX_JUMP_M_PER_S * budget_s or yaw_deg > MAX_JUMP_YAW_DEG_PER_S * budget_s:
                 logger.warning(
@@ -478,8 +434,7 @@ class RelocalizationModule(Module):
         self._last_fix_map_T_world = T
         self._last_fix_ts_s = now_s
 
-        # relocalize() returns map_T_world (scan_in_map = T @ scan_raw). We publish
-        # the world->map TF, so invert to world_T_map here.
+        # relocalize() returns map_T_world (scan_in_map = T @ scan_raw). We publish the world->map TF, so invert to world_T_map here.
         T_inv = np.linalg.inv(T)
         new_tf = Transform(
             translation=Vector3(*T_inv[:3, 3]),
