+++ dimos/perception/fiducial/marker_detection_stream_module.py
@@ -18,8 +18,7 @@ The module keeps the same transform chain used by offline marker tooling:
 quality-gated images, optional motion gating, marker fan-out, then one
 ``Detection3DArray`` per processed frame for LCM consumers.
 
-A second output, ``aggregated_detections``, carries one gated, robustly-aggregated
-pose per tag visit, with covariance and score. See ``AggregateTagBursts``.
+A second output, ``aggregated_detections``, carries one gated, robustly-aggregated pose per tag visit, with covariance and score. See ``AggregateTagBursts``.
 """
 
 from __future__ import annotations
@@ -71,8 +70,7 @@ class MarkerDetectionStreamModuleConfig(ModuleConfig):
     ambiguity_ratio_min: float = Field(
         2.0, ge=1.0
     )  # 1.0 = off; >1 drops mirror-ambiguous views. IPPE planar mirror-ambiguity, Collins & Bartoli 2014 https://link.springer.com/article/10.1007/s11263-014-0725-5
-    # Per-glimpse gates + aggregation for aggregated_detections; first pose only after
-    # min_observations * quality_window_s (1.0 s at defaults) -- the floor latency of a fix.
+    # per-glimpse gates + aggregation for aggregated_detections; first pose only after min_observations * quality_window_s (1.0 s at defaults) -- the floor latency of a fix
     aggregation: AggregationConfig = Field(default_factory=AggregationConfig)
 
 
@@ -121,8 +119,7 @@ class MarkerDetectionStreamModule(StreamModule[Image, Detection3DArray]):
                 )
             ),
         )
-        # tap yields every observation unchanged, so `detections` stays byte-identical and
-        # OpenCV runs once per frame -- a second .observable() would re-subscribe and re-detect.
+        # tap yields every observation unchanged, so `detections` stays byte-identical and OpenCV runs once per frame -- a second .observable() would re-subscribe and re-detect
         return markers.tap(self._aggregate).transform(
             MarkersPerFrame(frame_id=self.config.world_frame)
         )
